+++ Source/Veno/Audio/Oscillator/Helper/DetuneAlgos/VeNoXDetune.cpp
@@ -3,7 +3,7 @@
 
 void VeNoXDetune::prepareDetune (int voices)
 {
-    float detune = m_parameters->m_detuneDense->getValueForVoice(m_parameters->m_index);
+    double detune = m_parameters->m_detuneDense->getValueForVoice(m_parameters->m_index);
     if (m_lookup.empty () || voices != m_lastVoices || m_lastDetune != detune)
     {
         m_lastDetune = detune;
@@ -18,12 +18,12 @@ void VeNoXDetune::prepareDetune (int voices)
             return;
         }
 
-        float split = detune / (voices - 1);
-        float cents = 0;
+        double split = detune / (voices - 1);
+        double cents = 0;
         for (int i = 1; i < voices; ++i)
         {
             cents += split;
-            float plus;
+            double plus;
             if ((i & 1) == 1)
             {
                 plus = VeNo::Utils::centsToRatio (-cents);
@@ -40,7 +40,7 @@ void VeNoXDetune::prepareDetune (int voices)
 
 void VeNoXDetune::getRealDetune ()
 {
-    float d = m_lastDetune / m_detuneCents;
+    double d = m_lastDetune / m_detuneCents;
     m_currentDetune = std::atan (d * d * d * d) * m_detuneCents;
 }
