+++ Source/Veno/Audio/Oscillator/Modules/Limiter.cpp
@@ -7,9 +7,9 @@ Limiter::Limiter (const std::string& name, std::shared_ptr<OscillatorParameters>
     m_parameters = parameters;
 }
 
-void Limiter::apply (std::vector<float>& values, std::vector<float>& panned)
+void Limiter::apply (std::vector<double>& values, std::vector<double>& panned)
 {
-    for (float& value : values)
+    for (double& value : values)
     {
         value = VeNo::Utils::clamp (value, -1, 1);
     }
